reorganize JSON schema files
