+++ MANIFEST.in
@@ -3,4 +3,4 @@ graft artiq/examples
 include artiq/gui/logo*.svg
 include versioneer.py
 include artiq/_version.py
-include artiq/gateware/targets/kasli_generic.schema.json
+include artiq/coredevice/coredevice_generic.schema.json
