+++ artiq/coredevice/coredevice_generic.schema.json
